requirements.txt
fastapi
uvicorn
httpx
jinja2
python-dotenv
pydantic
pydantic-settings
python-multipart
pypdf
google-genai
psycopg2-binary
psycopg[binary]
pgvector
APScheduler
